pyproject.toml
[project]
# Defines project metadata.
name = "ct-rate-feature-benchmarks"
version = "0.1.0"
description = "Benchmarking classification and retrieval models on pre-computed CT-RATE visual and textual features."
authors = [
    { name = "Your Name", email = "your.email@example.com" },
]
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]

# Defines core project dependencies.
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.30.0", # For text feature processing and models
    "numpy>=1.20.0",
    "scikit-learn>=1.0", # For metrics and classical models
    "pandas>=1.5",       # For handling feature manifests
    "pyarrow>=10.0",     # For efficient feature loading (e.g., parquet)
    "hydra-core>=1.3.0", # For experiment configuration management
    "rich>=13.0.0",      # For better terminal logging
]

[project.optional-dependencies]
# Defines dependencies for running tests.
test = [
    "pytest>=8.0.0",
    "pytest-cov", # For code coverage analysis
]

# Defines dependencies for development and linting.
dev = [
    "black>=24.0.0",
    "ruff>=0.4.0",
    "pre-commit>=3.5.0",
]

[tool.pytest.ini_options]
# Configures the pytest runner.
pythonpath = ["src", "tests"]
testpaths = ["tests"]
addopts = "--cov=src --cov-report=term-missing"

[tool.ruff]
# Configures the Ruff linter and formatter.
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]

[tool.black]
# Configures the Black code formatter.
line-length = 88
target-version = ['py312']